requirements.txt
confluent-kafka
requests
psycopg2-binary
textblob
python-dotenv